Fix the javascript unit tests to use PhantomJS 2
The javascript unit tests could not be made for a while because of
an issue in the angular bitcore wallet client requiring phantomjs
version 2.
Since the new launcher was released a while ago already I
updated the karma phantom launcher and created some example tests to
prove we can create the tests again.
I tested loading the entire copayApp module in the unit tests to see if
everything was able to initialize. This worked, however the $log service
currently dumps the entire stacktrace of where the $log method was
called from in the console at the moment which creates a massive log
every startup. This does not occur for the chrome launcher so it has
something to do with the new phantomjs launcher. I don't think you want
to load the entire module for unit testing anyway so I don't see that as
a big problem as of now. It's nice that we can actually start testing
the code.